backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-socketio>=5.11.0
httpx>=0.27.0
pydantic>=2.6.0
anthropic>=0.40.0
openai>=1.50.0
google-genai>=1.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
jupyter>=1.0.0
ipykernel>=6.29.0
